projectile

+++ Assets/Scripts/Game Mechanics/Projectile.cs
@@ -4,6 +4,9 @@ using UnityEngine;
 
 public class Projectile : MonoBehaviour {
 	Rigidbody rigidbod;
+	[SerializeField] string createOnTimeOut;
+	[SerializeField] float timeOutTime;
+	[SerializeField] float destroyTime;
 	[HideInInspector] public int userId;
 	PhotonView photonView;
 	bool collided = false;
@@ -12,26 +15,26 @@ public class Projectile : MonoBehaviour {
 		object[] data = GetComponent<PhotonView>().instantiationData;
 		userId = (int)data[0];
 		rigidbod = gameObject.GetComponent<Rigidbody>();
-		if (photonView.isMine) {
-			StartCoroutine (fuseTime ());
+		if (photonView.isMine && timeOutTime > 0) {
+			StartCoroutine(timeOut ());
 		}
 	}
 
-	void blowUp() {
+	void createObject() {
 		gameObject.GetComponent<Renderer>().enabled = false;
 		rigidbod.constraints = RigidbodyConstraints.FreezePositionX | RigidbodyConstraints.FreezePositionY | RigidbodyConstraints.FreezePositionZ | RigidbodyConstraints.FreezeRotation;
-		PhotonNetwork.Instantiate("dynamiteExplosion", gameObject.transform.position, Quaternion.identity, 0, new object[] {userId});
-		StartCoroutine(destroyObject());
+		PhotonNetwork.Instantiate(createOnTimeOut, gameObject.transform.position, Quaternion.identity, 0, new object[] {userId});
+		if (destroyTime > 0) { StartCoroutine(destroyObject()); }
 	}
 
 
 	private IEnumerator destroyObject() {
-		yield return new WaitForSeconds(3f);
+		yield return new WaitForSeconds(destroyTime);
 		PhotonNetwork.Destroy(gameObject);
 	}
 
-	private IEnumerator fuseTime() {
-		yield return new WaitForSeconds (3f);
-		blowUp ();
+	private IEnumerator timeOut() {
+		yield return new WaitForSeconds (timeOutTime);
+		createObject();
 	}
 }
